Additional Sync Byte
Instead of just using the single SBUS header byte for syncing packets, I added storage of the previous byte and am making sure that the previous byte is one of the accepted SBUS footer values. This provides two bytes for syncing the start of the SBUS packet and improves robustness, especially in cases where the Teensy software is starts in the middle of a stream of SBUS data.