fixed Parse overwritting buf_
If Parse() returns true, the while loop will still run once after(I am guessing because there is still data in the serial RX buffer), overwritting the buffer.
This is a non issue under normal conditions, as it would just overwrite near identical data.
However if the second byte is equal to the header (0x0F), Parse() will assume that second byte to be the footer. 
This is because the execution branch at the end does not do "prev_byte_ = curr_byte_", while still having set curr_byte_ to the current package, hence skipping the header in the evaluation.